tax_records_spiders/requirements.txt
nameparser==1.0.4
usaddress-scourgify==0.1.9
pyyaml==3.13
awscli==1.16.277
PyPDF2==1.26.0
mysql-connector==2.2.9